+++ apps/client/src/pages/level/components/LevelScene.tsx
@@ -1,7 +1,6 @@
-// components/LevelScene.tsx
-import * as React from 'react';
 import { cn } from '@pinback/design-system/utils';
-import { type TreeLevel } from '../utils/treeLevel';
+import { TreeLevel } from '@pages/level/types/treeLevelType';
+import { HTMLAttributes } from 'react';
 
 import chippi_level1 from '../../../assets/chippi_level1.png';
 import chippi_level2 from '../../../assets/chippi_level2.png';
@@ -17,23 +16,18 @@ const SCENE_BY_LEVEL: Record<TreeLevel, string> = {
   5: chippi_level5,
 } as const;
 
-interface LevelSceneProps extends React.HTMLAttributes<HTMLDivElement> {
+interface LevelSceneProps extends HTMLAttributes<HTMLDivElement> {
   level: TreeLevel;
 }
 
-export default function LevelScene({
-  level,
-  className,
-  ...rest
-}: LevelSceneProps) {
+export default function LevelScene({ level, className }: LevelSceneProps) {
   const src = SCENE_BY_LEVEL[level];
   return (
-    <div className={cn('relative w-full', className)} {...rest}>
+    <div className={cn('absolute inset-0', className)} aria-hidden="true">
       <img
         src={src}
-        alt={`지식나무 숲 Lv.${level}`}
-        className="h-auto w-full select-none rounded-[1.2rem] object-cover"
         draggable={false}
+        className="pointer-events-none h-full w-full select-none rounded-[1.2rem] object-contain object-[right_bottom]"
       />
     </div>
   );
